organize directory structure; test running calicost in snakemake
