+++ config.yaml
@@ -21,5 +21,13 @@ chromosomes: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
 # input
 spaceranger_dir: "/u/congma/ragr-data/datasets/DingLab/HT112/U1/"
 
+# CalicoST parameters
+random_state: [0, 1, 2, 3, 4]
+hgtable_file: "/u/congma/ragr-data/users/congma/Codes/STARCH_crazydev/hgTables_hg38_gencode.txt"
+filtergenelist_file: "/n/fs/ragr-data/users/congma/references/cellranger_refdata-gex-GRCh38-2020-A/genes/ig_gene_list.txt"
+filterregion_file: "/n/fs/ragr-data/users/congma/references/cellranger_refdata-gex-GRCh38-2020-A/genes/HLA_regions.bed"
+tumorprop_file: "/u/congma/ragr-data/datasets/DingLab/HT112/HT112C1-T1_U1/ht112_u1_rctd_prop.tsv"
+
 # output
 outputdir: "/u/congma/ragr-data/users/congma/Datasets/test_calico"
+output_foldername: "calicost"
