+++ resources/template_configfile_joint
@@ -1,0 +1,43 @@
+input_filelist : /u/congma/ragr-data/datasets/DingLab/HT112/HT112_joint/bamfile_list.tsv
+snp_dir : /u/congma/ragr-data/datasets/DingLab/HT112/HT112_joint/visium_snpnew
+output_dir : /u/congma/ragr-data/datasets/DingLab/HT112/HT112_joint/cna_normalmixture_04
+
+# supporting files and preprocessing arguments
+geneticmap_file : /u/congma/ragr-data/users/congma/Codes/CalicoST/resources/genetic_map_GRCh38_merged.tab
+hgtable_file : /home/congma/congma/codes/locality-clustering-cnv/data/hgTables_hg38_gencode.txt
+normalidx_file : None
+tumorprop_file : None
+filtergenelist_file : /home/congma/congma/codes/locality-clustering-cnv/data/ig_gene_list.txt
+filterregion_file : /home/congma/congma/codes/locality-clustering-cnv/data/HLA_regions.bed
+avg_umi_perbinspot : 300
+bafonly : False
+
+# phase switch probability
+nu : 1.0
+logphase_shift : 0
+npart_phasing : 3
+
+# HMRF configurations
+n_clones : 3
+min_spots_per_clone : 50
+maxspots_pooling : 13
+tumorprop_threshold : 0.7
+max_iter_outer : 20
+nodepotential : weighted_sum
+initialization_method : rectangle
+num_hmrf_initialization_start : 0
+num_hmrf_initialization_end : 1
+
+# HMM configurations
+n_states : 7
+params : smp
+t : 1-1e-5
+fix_NB_dispersion : False
+shared_NB_dispersion : True
+fix_BB_dispersion : False
+shared_BB_dispersion : True
+max_iter : 30
+tol : 0.0001
+spatial_weight : 1.0
+gmm_random_state : 0
+np_threshold : 1.5
