+++ src/calicost/__init__.py
@@ -1,0 +1,1 @@
+__version__ = 'v1.0.0'
