+++ src/calicost/hmm_NB_sharedstates.py
