+++ src/calicost/hmm_gaussian.py
