+++ src/calicost/hmrf.py
