+++ src/calicost/utils_hmrf.py
@@ -1,25 +1,12 @@
-import logging
-from turtle import reset
 import numpy as np
 import pandas as pd
-from numba import njit
 import scipy.special
 import scipy.sparse
-from sklearn.mixture import GaussianMixture
-from sklearn.cluster import KMeans
-from sklearn.metrics import adjusted_rand_score, silhouette_score
 from sklearn.neighbors import kneighbors_graph
 from sklearn.neighbors import NearestNeighbors
-import networkx as nx
-from tqdm import trange
 import copy
-from pathlib import Path
-from hmm_NB_BB_phaseswitch import *
-from utils_distribution_fitting import *
-from utils_IO import *
-from simple_sctransform import *
-
-import warnings
+import anndata
+import scanpy as sc
 from statsmodels.tools.sm_exceptions import ValueWarning
